Build libmonoutils under none desktop Windows API family.

Commit migrated from https://github.com/mono/mono/commit/390697c8b2e2faf225440dc2b4c8ba73eb4dd85c

+++ src/mono/mono/utils/mono-logger.c
@@ -291,7 +291,7 @@ mono_trace_set_mask_string (const char *value)
 			continue;
 		}
 		for (i = 0; valid_flags[i]; i++) {
-			int len = strlen (valid_flags[i]);
+			size_t len = strlen (valid_flags[i]);
 			if (strncmp (tok, valid_flags[i], len) == 0 && (tok[len] == 0 || tok[len] == ',')) {
 				flags |= valid_masks[i];
 				tok += len;
